Exploded core into his parts
