+++ learnrl/agents/standard/__init__.py
@@ -4,7 +4,7 @@
 from learnrl.evaluation import QLearning
 from learnrl.control import Greedy
 from learnrl.estimators import TableEstimator
-from learnrl.core import Agent
+from learnrl import Agent
 from learnrl.memory import Memory
 
 from gym import spaces
