+++ learnrl/playground.py
@@ -8,148 +8,8 @@ import warnings
 from time import time
 from gym import Env
 
+from learnrl import Agent, TurnEnv
 from learnrl.callbacks import CallbackList, Logger
-
-
-class Agent():
-
-    """ A general structure for reinforcement learning agents    
-    
-    It uses by default a :class:`Memory`
-
-    Attributes
-    ----------
-
-        name: :class:`str`
-            The Agent's name
-        memory: :class:`Memory`
-            The Agent's memory
-    
-    """
-
-    def act(self, observation, greedy=False):
-        """ How the :ref:`Agent` act given an observation
-        
-        Parameters
-        ----------
-            observation:
-                The observation given by the |gym.Env|
-            greedy: bool
-                If True, act greedely (without exploration)
-
-        """
-        raise NotImplementedError
-
-    def learn(self):
-        """ How the :ref:`Agent` learns from his experiences 
-        
-        Returns
-        -------
-            logs: dict
-                The agent learning logs.
-
-        """
-        raise NotImplementedError
-
-    def remember(self, observation, action, reward, done, next_observation=None, info={}, **param):
-        """ Uses the agent's :class:`Memory` to remember experiences
-        
-        Often, the agent will use a |hash| to store observations efficiently
-
-        Example
-        -------
-            >>>  self.memory.remember(self.observation_encoder(observation),
-            ...                       self.action_encoder(action),
-            ...                       reward, done, 
-            ...                       self.observation_encoder(next_observation), 
-            ...                       info, **param)
-        """
-        raise NotImplementedError
-    
-
-class MultiEnv(Env):
-
-    r"""
-    A layer over the gym |gym.Env| class able to handle environements with multiple agents.
-   
-    .. note::
-        A :ref:`MultiEnv` must be in a :ref:`Playground` in order to work !
-
-    | The main add in MultiEnv is the method: 
-    |   turn
-
-    On top of the main API basic methodes (see |gym.Env|):
-        * step: take a step of the |gym.Env| given the action of the active player
-        * reset: reset the |gym.Env| and returns the first observation
-        * render
-        * close 
-        * seed
-
-    Attributes
-    ----------
-        action_space: |gym.Space|
-            The Space object corresponding to actions  
-        observation_space: |gym.Space|
-            The Space object corresponding to observations  
-        reward_range: :class:`tuple`
-            | A tuple corresponding to the min and max possible rewards.
-            | A default reward range set to [-inf,+inf] already exists. 
-            | Set it if you want a narrower range.
-
-    """
-
-    def step(self, action):
-        """Perform a step of the environement
-        
-        Parameters
-        ----------
-            action:
-                The action taken by the agent who's turn was given by :meth:`turn`.
-        
-        Return
-        ------
-            observation: 
-                The observation to give to the :class:`Agent`.
-            reward: :class:`float`
-                The reward given to the :class:`Agent` for this step.
-            done: :class:`bool`
-                Is the environement done after this step ?
-            info: :class:`dict`
-                Additional informations given by the |gym.Env|.
-            
-        """
-        raise NotImplementedError
-
-    def turn(self, state):
-        """Give the turn to the next agent to play
-    
-        Assuming that agents are represented by a list like range(n_player)
-        where n_player is the number of players in the game.
-        
-        Parameters
-        ----------
-            state: 
-                | The real state of the environement.
-                | Should be enough to determine which is the next agent to play.
-
-        Return
-        ------
-            agent_id: :class:`int`
-                The next player id
-
-        """
-        raise NotImplementedError
-
-    def reset(self):
-        """Reset the environement and returns the initial state
-
-        Return
-        ------
-            observation:
-                The observation for the first :class:`Agent` to play
-        
-        """
-        raise NotImplementedError
 
 
 class Playground():
@@ -158,10 +18,10 @@ class Playground():
 
     Attributes
     ----------
-        env: :class:`MultiEnv` or |gym.Env|
+        env: :class:`TurnEnv` or |gym.Env|
             Environement in which agents will play
         agents: list or :class:`Agent`
-            List of :class:`Agent` to run on the :class:`MultiEnv`
+            List of :class:`Agent` to run on the :class:`TurnEnv`
 
     """
 
@@ -185,7 +45,7 @@ class Playground():
             episodes: int
                 Number of episodes to run.
             render: bool
-                If True, call :meth:`MultiEnv.render` every step.
+                If True, call :meth:`TurnEnv.render` every step.
             learn: bool
                 If True, call :meth:`Agent.learn` every step.
             cycle_len: int
@@ -243,7 +103,7 @@ class Playground():
 
                 if render: self.env.render()
 
-                agent_id = self.env.turn(observation) if isinstance(self.env, MultiEnv) else 0
+                agent_id = self.env.turn(observation) if isinstance(self.env, TurnEnv) else 0
                 if agent_id >= len(previous):
                     raise ValueError(f'Not enough agents to play environement {self.env}')
                 agent = self.agents[agent_id]
